+++ ui/src/pages/chat/chat-pane-session.ts
@@ -1,15 +1,10 @@
 import type {
   SessionCatalogTranscriptItem,
   SessionsCatalogReadResult,
-  TaskSuggestion,
-  TaskSuggestionEvent,
-  TaskSuggestionsAcceptResult,
-  TaskSuggestionsListResult,
 } from "../../../../packages/gateway-protocol/src/index.js";
 import type { ControlUiSessionPullRequest } from "../../../../src/gateway/control-ui-contract.js";
 import type { GatewaySessionRow } from "../../api/types.ts";
 import { selectApplicationSession } from "../../app/agent-selection.ts";
-import { hasOperatorAdminAccess } from "../../app/operator-access.ts";
 import { clampText } from "../../lib/format.ts";
 import { isGatewayMethodAdvertised } from "../../lib/gateway-methods.ts";
 import { resolveSessionDisplayName } from "../../lib/session-display.ts";
@@ -27,11 +22,6 @@ import {
 } from "../../lib/sessions/catalog-key.ts";
 import { resolveSessionKey, scopedAgentParamsForSession } from "../../lib/sessions/index.ts";
 import { parseAgentSessionKey } from "../../lib/sessions/session-key.ts";
-import {
-  taskSuggestionAcceptParams,
-  type TaskSuggestionAcceptMode,
-} from "../../lib/task-suggestion-acceptance.ts";
-import { discoverCloudProfiles } from "../new-session/cloud-profile-discovery.ts";
 import { catalogMessageId } from "./catalog-message-id.ts";
 import { refreshChatAvatar } from "./chat-avatar.ts";
 import {
@@ -46,8 +36,8 @@ import {
   nativeHistoryMessageIdentity,
   summarizeSessionPullRequests,
 } from "./chat-pane-shared.ts";
-import { ChatPaneSharing } from "./chat-pane-sharing.ts";
 import { applySelectedSessionProjection } from "./chat-pane-state.ts";
+import { ChatPaneTaskSuggestions } from "./chat-pane-task-suggestions.ts";
 import { flushChatQueueForEvent } from "./chat-send-actions.ts";
 import { flushChatQueueAfterIdleSessionReconciliation } from "./chat-session.ts";
 import type { ChatPageHost } from "./chat-state-host.ts";
@@ -73,76 +63,7 @@ import {
 } from "./composer-persistence.ts";
 import { scheduleChatScroll } from "./scroll.ts";
 
-export abstract class ChatPaneSession extends ChatPaneSharing {
-  protected async ensureTaskSuggestionCloudProfiles(): Promise<void> {
-    const scope = this.captureConnectionScope();
-    if (
-      !scope ||
-      this.taskSuggestions.length === 0 ||
-      this.taskSuggestionCloudProfileGeneration === scope.generation ||
-      !hasOperatorAdminAccess(scope.context.gateway.snapshot.hello?.auth ?? null) ||
-      isGatewayMethodAdvertised(scope.context.gateway.snapshot, "taskSuggestions.accept") !== true
-    ) {
-      return;
-    }
-    // Profile metadata is connection-stable. Mark the generation before the
-    // request so repeated renders cannot turn this optional affordance into polling.
-    this.taskSuggestionCloudProfileGeneration = scope.generation;
-    if (isGatewayMethodAdvertised(scope.context.gateway.snapshot, "environments.list") !== true) {
-      return;
-    }
-    try {
-      const profiles = await discoverCloudProfiles(scope.client, true);
-      if (!this.isConnectionScopeCurrent(scope)) {
-        return;
-      }
-      this.taskSuggestionCloudProfiles = profiles.map((profile) => ({ id: profile.id }));
-      this.requestUpdate();
-    } catch {
-      // Cloud is optional; a failed one-shot discovery leaves the disabled hint.
-    }
-  }
-
-  protected async refreshTaskSuggestions(): Promise<void> {
-    const requestVersion = ++this.taskSuggestionsRequestVersion;
-    const scope = this.captureConnectionScope();
-    if (
-      !scope ||
-      !isGatewayMethodAdvertised(scope.context.gateway.snapshot, "taskSuggestions.list")
-    ) {
-      this.taskSuggestions = [];
-      this.requestUpdate();
-      return;
-    }
-    const sessionKey = scope.state.sessionKey;
-    if (parseCatalogSessionKey(sessionKey)) {
-      this.taskSuggestions = [];
-      this.requestUpdate();
-      return;
-    }
-    const agentId = resolveChatAgentId(scope.state);
-    try {
-      const result = await scope.client.request<TaskSuggestionsListResult>("taskSuggestions.list", {
-        agentId,
-      });
-      if (
-        requestVersion !== this.taskSuggestionsRequestVersion ||
-        !this.isConnectionScopeCurrent(scope) ||
-        sessionKey !== scope.state.sessionKey
-      ) {
-        return;
-      }
-      this.taskSuggestions = result.suggestions.filter((suggestion) =>
-        this.suggestionMatchesCurrentSession(suggestion),
-      );
-      this.requestUpdate();
-    } catch {
-      // Suggestions are an optional ephemeral affordance; chat remains usable
-      // when an older Gateway or a reconnect loses the process-local registry.
-      // Keep event-delivered cards when a background reconciliation fails.
-    }
-  }
-
+export abstract class ChatPaneSession extends ChatPaneTaskSuggestions {
   protected async refreshSessionPullRequests(options: { refresh?: boolean } = {}): Promise<void> {
     const scope = this.captureConnectionScope();
     if (
@@ -223,92 +144,6 @@ export abstract class ChatPaneSession extends ChatPaneSharing {
     this.dismissedSessionPullRequestIds = dismissChatPullRequest(sessionKey, pullRequest);
     this.requestUpdate();
   };
-
-  protected handleTaskSuggestionEvent(event: TaskSuggestionEvent): void {
-    if (event.action === "created") {
-      if (!this.suggestionMatchesCurrentSession(event.suggestion)) {
-        return;
-      }
-      this.taskSuggestions = [
-        event.suggestion,
-        ...this.taskSuggestions.filter((item) => item.id !== event.suggestion.id),
-      ];
-    } else {
-      this.taskSuggestions = this.taskSuggestions.filter((item) => item.id !== event.taskId);
-      this.taskSuggestionBusyIds.delete(event.taskId);
-    }
-    this.requestUpdate();
-    // The replacement snapshot includes the event plus unrelated suggestions;
-    // its request version prevents any older snapshot from overwriting either.
-    void this.refreshTaskSuggestions();
-  }
-
-  protected readonly acceptTaskSuggestion = (
-    suggestion: TaskSuggestion,
-    mode: TaskSuggestionAcceptMode,
-    cloudProfileId?: string,
-  ): Promise<void> => this.resolveTaskSuggestion(suggestion, "accept", mode, cloudProfileId);
-
-  protected readonly dismissTaskSuggestion = (suggestion: TaskSuggestion): Promise<void> =>
-    this.resolveTaskSuggestion(suggestion, "dismiss");
-
-  protected async resolveTaskSuggestion(
-    suggestion: TaskSuggestion,
-    action: "accept" | "dismiss",
-    mode: TaskSuggestionAcceptMode = "worktree",
-    cloudProfileId?: string,
-  ): Promise<void> {
-    const scope = this.captureConnectionScope();
-    if (
-      !scope ||
-      !this.suggestionMatchesCurrentSession(suggestion) ||
-      this.taskSuggestionOperations.has(suggestion.id)
-    ) {
-      return;
-    }
-    const sessionKey = scope.state.sessionKey;
-    const operation = Symbol();
-    const isCurrent = () =>
-      this.isConnectionScopeCurrent(scope) &&
-      scope.state.sessionKey === sessionKey &&
-      this.taskSuggestionOperations.get(suggestion.id) === operation;
-    this.taskSuggestionOperations.set(suggestion.id, operation);
-    this.taskSuggestionBusyIds.add(suggestion.id);
-    this.requestUpdate();
-    try {
-      let acceptedKey: string | undefined;
-      if (action === "accept") {
-        const result = await scope.client.request<TaskSuggestionsAcceptResult>(
-          "taskSuggestions.accept",
-          taskSuggestionAcceptParams(suggestion.id, mode, cloudProfileId),
-        );
-        acceptedKey = result.key;
-      } else {
-        await scope.client.request("taskSuggestions.dismiss", { taskId: suggestion.id });
-      }
-      if (!isCurrent()) {
-        return;
-      }
-      this.taskSuggestions = this.taskSuggestions.filter((item) => item.id !== suggestion.id);
-      if (acceptedKey && mode !== "session") {
-        this.onPaneSessionChange?.(this.paneId, acceptedKey);
-      }
-    } catch (error) {
-      if (!isCurrent()) {
-        return;
-      }
-      scope.state.lastError = error instanceof Error ? error.message : String(error);
-      scope.state.chatError = scope.state.lastError;
-    } finally {
-      if (this.taskSuggestionOperations.get(suggestion.id) === operation) {
-        this.taskSuggestionOperations.delete(suggestion.id);
-        this.taskSuggestionBusyIds.delete(suggestion.id);
-        if (this.isConnectionScopeCurrent(scope) && scope.state.sessionKey === sessionKey) {
-          this.requestUpdate();
-        }
-      }
-    }
-  }
 
   protected deferSessionHydrationUntilTranscript(
     sessionKey: string,
