+++ ui/src/pages/chat/components/chat-task-suggestions.ts
@@ -4,11 +4,43 @@ import type { TaskSuggestion } from "../../../../../packages/gateway-protocol/sr
 import { icons } from "../../../components/icons.ts";
 import "../../../components/web-awesome.ts";
 import { t } from "../../../i18n/index.ts";
-import { copyToClipboard } from "../../../lib/clipboard.ts";
 import { repoName } from "../../../lib/session-display.ts";
 import type { TaskSuggestionAcceptMode } from "../../../lib/task-suggestion-acceptance.ts";
 
 type TaskSuggestionCloudProfile = { id: string };
+
+export type ChatTaskSuggestionTrayProps = {
+  taskSuggestions?: TaskSuggestion[];
+  taskSuggestionBusyIds?: ReadonlySet<string>;
+  taskSuggestionCloudProfiles?: TaskSuggestionCloudProfile[];
+  taskSuggestionCopiedIds?: ReadonlySet<string>;
+  onCopyTaskSuggestionPrompt?: (suggestion: TaskSuggestion) => void;
+  canAcceptTaskSuggestions?: boolean;
+  canAcceptTaskSuggestionModes?: boolean;
+  canDismissTaskSuggestions?: boolean;
+  onAcceptTaskSuggestion?: (
+    suggestion: TaskSuggestion,
+    mode: TaskSuggestionAcceptMode,
+    cloudProfileId?: string,
+  ) => void;
+  onDismissTaskSuggestion?: (suggestion: TaskSuggestion) => void;
+};
+
+export function renderChatTaskSuggestionTray(props: ChatTaskSuggestionTrayProps) {
+  return renderChatTaskSuggestions({
+    suggestions: props.taskSuggestions ?? [],
+    busyIds: props.taskSuggestionBusyIds ?? new Set(),
+    cloudProfiles: props.taskSuggestionCloudProfiles ?? [],
+    copiedIds: props.taskSuggestionCopiedIds ?? new Set(),
+    onCopyPrompt: (suggestion) => props.onCopyTaskSuggestionPrompt?.(suggestion),
+    canAccept: props.canAcceptTaskSuggestions === true,
+    canAcceptModes: props.canAcceptTaskSuggestionModes === true,
+    canDismiss: props.canDismissTaskSuggestions === true,
+    onAccept: (suggestion, mode, cloudProfileId) =>
+      props.onAcceptTaskSuggestion?.(suggestion, mode, cloudProfileId),
+    onDismiss: (suggestion) => props.onDismissTaskSuggestion?.(suggestion),
+  });
+}
 
 // Mirrors the TUI sanitizer to prevent directionality spoofing. This stays local
 // because the Control UI cannot import core src/ modules.
@@ -28,6 +60,8 @@ export function renderChatTaskSuggestions(props: {
     cloudProfileId?: string,
   ) => void;
   onDismiss: (suggestion: TaskSuggestion) => void;
+  onCopyPrompt: (suggestion: TaskSuggestion) => void;
+  copiedIds: ReadonlySet<string>;
   canAcceptModes: boolean;
 }) {
   if (props.suggestions.length === 0 || (!props.canAccept && !props.canDismiss)) {
@@ -96,80 +130,81 @@ export function renderChatTaskSuggestions(props: {
                     ? t("chat.taskSuggestions.starting")
                     : t("chat.taskSuggestions.startWorktree")}
                 </button>
-                ${props.canAcceptModes
-                  ? html`
-                      <wa-dropdown
-                        class="task-suggestion__menu"
-                        placement="bottom-end"
-                        @wa-select=${(
-                          event: CustomEvent<{ item: HTMLElement & { value?: string } }>,
-                        ) => {
-                          const item = event.detail.item;
-                          if (item.value === "local") {
-                            accept("local");
-                          } else if (item.value === "session") {
-                            accept("session");
-                          } else if (item.value === "cloud") {
-                            const profileId = item.dataset.cloudProfile;
-                            if (profileId) {
-                              accept("cloud", profileId);
-                            }
-                          } else if (item.value === "copy-prompt") {
-                            // Copies the raw prompt: fidelity matters at the paste
-                            // target; sanitizing is a display-only concern.
-                            void copyToClipboard(suggestion.prompt);
-                          }
-                        }}
-                      >
-                        <button
-                          slot="trigger"
-                          class="btn task-suggestion__menu-trigger"
-                          type="button"
-                          ?disabled=${busy || !props.canAccept}
-                          title=${props.canAccept ? "" : t("chat.taskSuggestions.adminRequired")}
-                          aria-label=${t("chat.taskSuggestions.moreActions")}
-                          aria-haspopup="menu"
-                          aria-expanded="false"
-                        >
-                          ${icons.chevronDown}
-                        </button>
-                        <wa-dropdown-item value="local" ?disabled=${busy || !props.canAccept}>
-                          ${t("chat.taskSuggestions.startLocal")}
-                        </wa-dropdown-item>
-                        ${cloudProfiles.length === 0
-                          ? html`
-                              <wa-dropdown-item
-                                value="cloud"
-                                disabled
-                                title=${t("chat.taskSuggestions.noCloudConfigured")}
-                              >
-                                ${t("chat.taskSuggestions.startCloudGeneric")}
-                              </wa-dropdown-item>
-                            `
-                          : cloudProfiles.map(
-                              (profile) => html`
+                ${html`
+                  <wa-dropdown
+                    class="task-suggestion__menu"
+                    placement="bottom-end"
+                    @wa-select=${(
+                      event: CustomEvent<{ item: HTMLElement & { value?: string } }>,
+                    ) => {
+                      const item = event.detail.item;
+                      if (item.value === "local") {
+                        accept("local");
+                      } else if (item.value === "session") {
+                        accept("session");
+                      } else if (item.value === "cloud") {
+                        const profileId = item.dataset.cloudProfile;
+                        if (profileId) {
+                          accept("cloud", profileId);
+                        }
+                      } else if (item.value === "copy-prompt") {
+                        props.onCopyPrompt(suggestion);
+                      }
+                    }}
+                  >
+                    <button
+                      slot="trigger"
+                      class="btn task-suggestion__menu-trigger"
+                      type="button"
+                      ?disabled=${busy}
+                      aria-label=${t("chat.taskSuggestions.moreActions")}
+                      aria-haspopup="menu"
+                      aria-expanded="false"
+                    >
+                      ${icons.chevronDown}
+                    </button>
+                    ${props.canAcceptModes
+                      ? html`
+                          <wa-dropdown-item value="local" ?disabled=${busy || !props.canAccept}>
+                            ${t("chat.taskSuggestions.startLocal")}
+                          </wa-dropdown-item>
+                          ${cloudProfiles.length === 0
+                            ? html`
                                 <wa-dropdown-item
                                   value="cloud"
-                                  data-cloud-profile=${profile.id}
-                                  ?disabled=${busy || !props.canAccept}
+                                  disabled
+                                  title=${t("chat.taskSuggestions.noCloudConfigured")}
                                 >
-                                  ${cloudProfiles.length > 1
-                                    ? t("chat.taskSuggestions.startCloud", {
-                                        profile: profile.label,
-                                      })
-                                    : t("chat.taskSuggestions.startCloudGeneric")}
+                                  ${t("chat.taskSuggestions.startCloudGeneric")}
                                 </wa-dropdown-item>
-                              `,
-                            )}
-                        <wa-dropdown-item value="session" ?disabled=${busy || !props.canAccept}>
-                          ${t("chat.taskSuggestions.fixInSession")}
-                        </wa-dropdown-item>
-                        <wa-dropdown-item value="copy-prompt">
-                          ${t("chat.taskSuggestions.copyPrompt")}
-                        </wa-dropdown-item>
-                      </wa-dropdown>
-                    `
-                  : nothing}
+                              `
+                            : cloudProfiles.map(
+                                (profile) => html`
+                                  <wa-dropdown-item
+                                    value="cloud"
+                                    data-cloud-profile=${profile.id}
+                                    ?disabled=${busy || !props.canAccept}
+                                  >
+                                    ${cloudProfiles.length > 1
+                                      ? t("chat.taskSuggestions.startCloud", {
+                                          profile: profile.label,
+                                        })
+                                      : t("chat.taskSuggestions.startCloudGeneric")}
+                                  </wa-dropdown-item>
+                                `,
+                              )}
+                          <wa-dropdown-item value="session" ?disabled=${busy || !props.canAccept}>
+                            ${t("chat.taskSuggestions.fixInSession")}
+                          </wa-dropdown-item>
+                        `
+                      : nothing}
+                    <wa-dropdown-item value="copy-prompt">
+                      ${props.copiedIds.has(suggestion.id)
+                        ? t("chat.taskSuggestions.promptCopied")
+                        : t("chat.taskSuggestions.copyPrompt")}
+                    </wa-dropdown-item>
+                  </wa-dropdown>
+                `}
               </div>
             </div>
           </article>
